Update directory structure
